requirements.txt
dipy==1.7.0
h5py==2.10.0
matplotlib==3.5.3
nibabel==3.2.2
numpy==1.18.5
pandas==1.3.5
scikit_image==0.19.3
scikit_learn==1.0.2
tensorflow_addons==0.14.0
tensorflow_gpu==2.3.0
 
  